Changes: Moved the Serial functions
Moved the Serial functions to Serial.js.
Moved button and baudrate selector to UI directory
